refactor: 오답 카드를 공용 컴포넌트로 이동하고 연습 문제 카드에 재사용
